+++ connect4/DiscType.cs
@@ -1,0 +1,10 @@
+namespace LineUp
+{
+    public enum DiscType
+    {
+        Ordinary,
+        Boring,
+        Magnetic
+    }
+}
+
