+++ connect4/GameMode.cs
@@ -1,0 +1,9 @@
+namespace LineUp
+{
+    public enum GameMode
+    {
+        HumanVsHuman,
+        HumanVsComputer
+    }
+}
+
