+++ connect4/PlayerId.cs
@@ -1,0 +1,9 @@
+namespace LineUp
+{
+    public enum PlayerId
+    {
+        One = 0,
+        Two = 1
+    }
+}
+
