+++ connect4/PlayerState.cs
@@ -1,0 +1,12 @@
+namespace LineUp
+{
+    public class PlayerState
+    {
+        public PlayerId Id { get; set; }
+        public bool IsComputer { get; set; }
+        public int Ordinary { get; set; }
+        public int Boring { get; set; }
+        public int Magnetic { get; set; }
+    }
+}
+
